separation of an interface from its corresponding class
