+++ src/GiveCRM.Web/GiveCRM.Web.csproj
@@ -118,6 +118,7 @@
     <Compile Include="Services\IMemberService.cs" />
     <Compile Include="Services\ISearchService.cs" />
     <Compile Include="Services\MailingListService.cs" />
+    <Compile Include="Services\MemberSearchFilterService.cs" />
     <Compile Include="Services\MemberService.cs" />
     <Compile Include="Services\SearchService.cs" />
     <Compile Include="Views\MenuHelper.cs" />
